+++ CHANGELOG.md
@@ -12,6 +12,7 @@ CHANGELOG
 - Add support for p6-b300 instances for all OSs except AL2.
 - Add alarm on missing clustermgtd heartbeat.
 - Support updates of `Tags` during cluster-updates.
+- Add `LaunchSpecificationOverrides` to cluster config to allow network interfaces to be customized by overriding the launch template of a compute resource.
 
 **BUG FIXES**
 - Add validation to block updates that change tag order. Blocking such change prevents update failures.